requirements.txt
# Previously this was necessary:
# pip uninstall torch -y && pip install torch --index-url https://download.pytorch.org/whl/cu128
# Now, it doesn't and I needed this:
# pip install --force-reinstall torch==2.7.0 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
# pip install --force-reinstall pandas scipy
numpy==1.26.4
tiktoken>=0.4.0
datasets>=2.11.0
smart-open>=6.3.0
boto3>=1.26.119
dotenv
torchtune
torchao
torch-optimi
torchvision
torchaudio
neptune-client
lm_eval
flash-attn
sentence_transformers